[KYUUBI #4757][UI] Move Statistics to Management Menu
