+++ kyuubi-server/web-ui/src/router/index.ts
@@ -20,9 +20,7 @@ import overviewRoutes from './overview'
 import workloadRoutes from './workload'
 import operationRoutes from './operation'
 import contactRoutes from './contact'
-import sessionRoutes from './session'
-import engineRoutes from './engine'
-import serverRoutes from './server'
+import managementRoutes from './management'
 
 const routes = [
   {
@@ -39,11 +37,9 @@ const routes = [
     redirect: 'overview',
     children: [
       ...overviewRoutes,
-      ...sessionRoutes,
       ...workloadRoutes,
       ...operationRoutes,
-      ...engineRoutes,
-      ...serverRoutes,
+      ...managementRoutes,
       ...contactRoutes
     ]
   }
