+++ kyuubi-server/web-ui/src/views/management/operation-statistics/index.vue
