+++ kyuubi-server/web-ui/src/views/management/server-statistics/index.vue
