+++ kyuubi-server/web-ui/src/views/management/session-statistics/index.vue
